# samples/openapi/step7-contract.yaml
...
openapi: 3.0.3
info:
  title: Payment API
  version: 1.0.0
security:
  - bearerAuth: []
paths:
...
  /users/{id}:
    get:
      operationId: getUserById
      summary: Get user by id
      description: Inherits the global bearerAuth requirement (no operation-level security).
      parameters:
        - name: id
          in: path
          required: true
          description: User id
          schema:
            type: string
        - name: verbose
          in: query
          required: false
          schema:
            type: string
            enum:
              - summary
              - full
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://auth.example.com/token
          scopes:
            admin: Full admin access
            user: Read-only user access
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
...